+++ src/get_endpoint_func.jl
@@ -1,13 +1,14 @@
 
 function get_endpoint(
     theta_init::Vector{Float64},
-    scan_func::Function, # h(theta) function for predictions or parameter
-    loss_func::Function,
+    #scan_func::Function, # h(theta) function for predictions or parameter
+    #loss_func::Function,
+    scan_loss_func::Function, # function returns tuple of scan value, loss value: (h, lambda)
     method::Symbol, # method::Val{:CICO_ONE_PASS},
     direction::Symbol = :right;
 
     loss_crit::Float64 = 0.0,
-    scan_scale::Symbol = :direct, # scale for scan_func XXX: not works
+    scan_scale::Symbol = :direct, # scale for scan_value XXX: not works
     # :direct, :log, :logit
     scale::Vector{Symbol} = fill(:direct, length(theta_init)),
     theta_bounds::Vector{Tuple{Float64,Float64}} = unscaling.(
@@ -31,15 +32,8 @@ function get_endpoint(
     if any(theta_init_outside_theta_bounds)
         throw(ArgumentError("theta_init is outside theta_bound: $(findall(theta_init_outside_theta_bounds))"))
     end
-    # scan_func initial should be within scan_bound
-    # XXX: maybe it is not necessary
-    #=
-    scan_func_init = scan_func(theta_init)
-    if (scan_func_init >= scan_bound && !isLeft) || scan_func_init <= scan_bound && isLeft)
-        throw(ArgumentError("init scan_func($theta_init) is outside of the scan_bound $scan_bound"))
-    end
-    =#
-    # 0 <= theta_bound[1] for :log
+
+    # 0 <= theta_bound for :log
     less_than_zero_theta_bounds = (scale .== :log) .& [theta_bounds[i][1] < 0 for i in 1:length(theta_init)]
     if any(less_than_zero_theta_bounds)
         throw(ArgumentError(":log scaled theta_bound min is negative: $(findall(less_than_zero_theta_bounds))"))
@@ -49,6 +43,14 @@ function get_endpoint(
     if any(less_than_zero_theta_bounds)
         throw(ArgumentError(":logit scaled theta_bound min is outside range [0,1]: $(findall(less_than_zero_theta_bounds))"))
     end
+    # scan_initial should be within scan_bound
+    # XXX: maybe it is not necessary
+    #=
+    scan_func_init = scan_func(theta_init)
+    if (scan_func_init >= scan_bound && !isLeft) || scan_func_init <= scan_bound && isLeft)
+        throw(ArgumentError("init scan_func($theta_init) is outside of the scan_bound $scan_bound"))
+    end
+    =#
     # loss_func(theta_init) < loss_crit
     # XXX: maybe it is not necessary
     #=
@@ -63,26 +65,22 @@ function get_endpoint(
     # transforming theta
     theta_init_gd = scaling.(theta_init, scale)
     # transforming scan fun
-    function scan_func_gd(theta_gd::Vector{Float64})
+    function scan_loss_func_gd(theta_gd::Vector{Float64})
         theta_g = copy(theta_gd)
         theta = unscaling.(theta_g, scale)
-        scan_value = scan_func(theta)
+        (scan_val, loss_val) = scan_loss_func(theta)
         # TODO: supreme value here
         # TODO: transformed by scan_scale
 
-        return isLeft ? (-1)*scan_value : scan_value
-    end
-
-    function loss_func_gd(theta_gd::Vector{Float64})
-        theta_g = copy(theta_gd)
-        theta = unscaling.(theta_g, scale)
-        # calculate function
-        loss_norm = loss_func(theta) - loss_crit
-
         # update counter
         counter += 1
-        return loss_norm
+
+        (
+            isLeft ? (-1)*scan_val : scan_val,
+            loss_val - loss_crit
+        )
     end
+
     theta_bounds_gd = scaling.(theta_bounds, scale)
 
     # TODO: transformed by scan_scale: scaling(scan_bound, scan_scale)
@@ -92,8 +90,7 @@ function get_endpoint(
     # calculate endpoint using base method
     (optf_gd, pp_gd, status) = get_right_endpoint(
         theta_init_gd,
-        scan_func_gd,
-        loss_func_gd,
+        scan_loss_func_gd,
         Val(method);
         theta_bounds = theta_bounds_gd,
         scan_bound = scan_bound_gd,
@@ -117,8 +114,7 @@ function get_endpoint(
     pps = [ temp_fun(pp_gd[i]) for i in 1:length(pp_gd) ]
     # transforming supreme back
     # if (isLeft && typeof(supreme_gd)!==Nothing) supreme_gd *= -1 end # change direction
-    # supreme = unscaling(supreme_gd, scan_scale)
-    supreme = 0
+    supreme = unscaling(supreme_gd, scan_scale)
 
     if (isLeft && typeof(optf_gd)!==Nothing) optf_gd *= -1 end # change direction
     # optf = unscaling(optf_gd, scan_scale)